# valo-mapper-api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateFolderRequest:
    properties:
      name:
        type: string
      parentFolderId:
        type: integer
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
...
  handlers.UpdateFolderRequest:
    properties:
      name:
        type: string
      parentFolderId:
        type: integer
    type: object
  handlers.UpdateLobbyRequest:
    properties:
      canvasState:
        $ref: '#/definitions/models.FullCanvasState'
    type: object
...
  models.CanvasAbility:
    properties:
      action:
        type: string
      currentLength:
        type: number
      currentPath:
        items:
          $ref: '#/definitions/models.Position'
        type: array
      currentRotation:
        type: number
      id:
        type: string
      isAlly:
        type: boolean
      name:
        type: string
      x:
        type: number
      "y":
        type: number
    type: object
  models.CanvasAgent:
    properties:
      id:
        type: string
      isAlly:
        type: boolean
      name:
        type: string
      role:
        type: string
      x:
        type: number
      "y":
        type: number
    type: object
  models.CanvasConnectingLine:
    properties:
      fromId:
        type: string
      id:
        type: string
      notes:
        type: string
      strokeColor:
        type: string
      strokeWidth:
        type: number
      toId:
        type: string
      uploadedImages:
        items:
          type: string
        type: array
      youtubeLink:
        type: string
    type: object
  models.CanvasDrawLine:
    properties:
      color:
        type: string
      id:
        type: string
      isArrowHead:
        type: boolean
      isDashed:
        type: boolean
      points:
        items:
          $ref: '#/definitions/models.Position'
        type: array
      size:
        type: number
      tool:
        type: string
    type: object
  models.CanvasImage:
    properties:
      height:
        type: number
      id:
        type: string
      src:
        type: string
      width:
        type: number
      x:
        type: number
      "y":
        type: number
    type: object
  models.CanvasText:
    properties:
      height:
        type: number
      id:
        type: string
      text:
        type: string
      width:
        type: number
      x:
        type: number
      "y":
        type: number
    type: object
  models.CanvasToolIcon:
    properties:
      height:
        type: number
      id:
        type: string
      width:
        type: number
      x:
        type: number
      "y":
        type: number
    type: object
  models.Folder:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      parentFolderId:
        type: integer
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  models.FullCanvasState:
    properties:
      abilitiesSettings:
        $ref: '#/definitions/models.IconSettings'
      agentsSettings:
        $ref: '#/definitions/models.IconSettings'
      currentPhaseIndex:
        type: integer
      editedPhases:
        items:
          type: integer
        type: array
      mapSide:
        type: string
      phases:
        items:
          $ref: '#/definitions/models.PhaseState'
        type: array
      selectedMap:
        $ref: '#/definitions/models.MapOption'
    type: object
  models.IconSettings:
    properties:
      allyColor:
        type: string
      borderOpacity:
        type: number
      borderWidth:
        type: integer
      enemyColor:
        type: string
      radius:
        type: integer
      scale:
        type: integer
    type: object
  models.Lobby:
    properties:
      canvasState:
        $ref: '#/definitions/models.FullCanvasState'
      createdAt:
        type: string
      lobbyCode:
        type: string
      mapSide:
        type: string
      selectedMapId:
        type: string
      updatedAt:
        type: string
    type: object
  models.MapOption:
    properties:
      id:
        type: string
      text:
        type: string
      textColor:
        type: string
    type: object
  models.PhaseState:
    properties:
      abilitiesOnCanvas:
        items:
          $ref: '#/definitions/models.CanvasAbility'
        type: array
      agentsOnCanvas:
        items:
          $ref: '#/definitions/models.CanvasAgent'
        type: array
      connectingLines:
        items:
          $ref: '#/definitions/models.CanvasConnectingLine'
        type: array
      drawLines:
        items:
          $ref: '#/definitions/models.CanvasDrawLine'
        type: array
      imagesOnCanvas:
        items:
          $ref: '#/definitions/models.CanvasImage'
        type: array
      textsOnCanvas:
        items:
          $ref: '#/definitions/models.CanvasText'
        type: array
      toolIconsOnCanvas:
        items:
          $ref: '#/definitions/models.CanvasToolIcon'
        type: array
    type: object
  models.Position:
    properties:
      x:
        type: number
      "y":
        type: number
    type: object
...
info:
  contact: {}
  description: API for ValoMapper backend services.
  title: ValoMapper API
  version: "1.0"
paths:
...
  /api/folders:
    get:
      description: Lists folders for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Folder'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List folders
      tags:
      - folders
    post:
      consumes:
      - application/json
      description: Creates a folder for the authenticated user.
      parameters:
      - description: Create folder request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateFolderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Folder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create folder
      tags:
      - folders
  /api/folders/{id}:
    delete:
      description: Deletes a folder owned by the authenticated user.
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete folder
      tags:
      - folders
    patch:
      consumes:
      - application/json
      description: Updates folder name or parent for the authenticated user.
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update folder request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Folder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update folder
      tags:
      - folders
  /api/lobbies:
    post:
      description: Creates a new lobby initialized with default map and empty phases.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Lobby'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create lobby
      tags:
      - lobbies
  /api/lobbies/{code}:
    get:
      description: Retrieves lobby and full canvas phases by lobby code.
      parameters:
      - description: Lobby code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lobby'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get lobby
      tags:
      - lobbies
    patch:
      consumes:
      - application/json
      description: Updates lobby canvas state by lobby code.
      parameters:
      - description: Lobby code
        in: path
        name: code
        required: true
        type: string
      - description: Canvas update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateLobbyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lobby'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update lobby canvas state
      tags:
      - lobbies
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Firebase bearer token in the form "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
  InternalAPIKey:
    description: Internal service API key.
    in: header
    name: X-Internal-API-Key
    type: apiKey
swagger: "2.0"